pyproject.toml
[project]
name = "l2m-cot"
version = "0.1.0"
description = "L2M-CoT Project"
authors = [
    {name = "User",email = "user@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai (>=2.11.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
